+++ src/structlog/processors.py
@@ -319,9 +319,9 @@ class JSONRenderer:
         serializer:
             A :func:`json.dumps`-compatible callable that will be used to
             format the string.  This can be used to use alternative JSON
-            encoders (default: :func:`json.dumps`).
-
-            .. seealso:: :doc:`performance` for examples.
+            encoders like `orjson <https://pypi.org/project/orjson/>`__ or
+            `RapidJSON <https://pypi.org/project/python-rapidjson/>`_
+            (default: :func:`json.dumps`).
 
     .. versionadded:: 0.2.0 Support for ``__structlog__`` serialization method.
     .. versionadded:: 15.4.0 *serializer* parameter.
@@ -334,8 +334,7 @@ class JSONRenderer:
         serializer: Callable[..., str | bytes] = json.dumps,
         **dumps_kw: Any,
     ) -> None:
-        if dumps_kw != {}:
-            dumps_kw.setdefault("default", _json_fallback_handler)
+        dumps_kw.setdefault("default", _json_fallback_handler)
         self._dumps_kw = dumps_kw
         self._dumps = serializer
 
@@ -345,9 +344,7 @@ class JSONRenderer:
         """
         The return type of this depends on the return type of self._dumps.
         """
-        if self._dumps_kw:
-            return self._dumps(event_dict, **self._dumps_kw)
-        return self._dumps(event_dict)
+        return self._dumps(event_dict, **self._dumps_kw)
 
 
 def _json_fallback_handler(obj: Any) -> Any:
@@ -916,7 +913,7 @@ class EventRenamer:
         return event_dict
 
 
-class AddCallingNamespace:
+class CallsiteNamespaceAddr:
     """
     Attempt to identify and add the caller namespace to the event dict
     under the ``namespace`` key.